create separete file for documentCard Widget ...widgets/document_card.dart
