+++ lib/screens/home_page.dart
@@ -3,7 +3,6 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../providers/auth_state_provider.dart';
 // Note: Adjusted import path for document_upload_page.dart as it's now in lib/screens/
 import '../widget/homepage_menu_bar_widget.dart';
-import '../models/document.dart';
 import '../providers/documents_provider.dart';
 import 'package:go_router/go_router.dart';
 
@@ -46,9 +45,6 @@ class _HomePageState extends ConsumerState<HomePage> {
     return Scaffold(
       appBar: AppBar(
         title: const Text('DocuBox'),
-        backgroundColor: Colors.blue.shade700,
-        foregroundColor: Colors.white,
-        elevation: 2,
         actions: [
           IconButton(
             onPressed: () {
@@ -73,7 +69,7 @@ class _HomePageState extends ConsumerState<HomePage> {
               controller: _searchController,
               decoration: InputDecoration(
                 hintText: 'Search documents...',
-                prefixIcon: const Icon(Icons.search, color: Colors.white70),
+                prefixIcon: const Icon(Icons.search,),
                 suffixIcon:
                     _searchController.text.isNotEmpty
                         ? IconButton(
@@ -205,109 +201,4 @@ class _HomePageState extends ConsumerState<HomePage> {
     );
   }
 }
-
-class DocumentCard extends StatelessWidget {
-  final Document document;
-
-  const DocumentCard({super.key, required this.document});
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      margin: const EdgeInsets.only(bottom: 16.0),
-      elevation: 3,
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-      child: ConstrainedBox(
-        constraints: const BoxConstraints(minHeight: 120),
-        child: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                children: [
-                  Icon(
-                    Icons.description,
-                    color: Theme.of(context).colorScheme.primary,
-                  ),
-                  const SizedBox(width: 12),
-                  Expanded(
-                    child: Text(
-                      document.name.isNotEmpty
-                          ? document.name
-                          : 'Untitled Document',
-                      style: const TextStyle(
-                        fontSize: 18,
-                        fontWeight: FontWeight.bold,
-                      ),
-                  Icon(icon:Icons.description,),
-                  const SizedBox(width: 12),
-                  Expanded(
-                    child: Text(
-                      document.name.isNotEmpty ? document.name : 'Untitled Document',
-                      style: const TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  ),
-                  IconButton(
-                    icon: const Icon(Icons.more_vert),
-                    onPressed: () {
-                      ScaffoldMessenger.of(context).showSnackBar(
-                        const SnackBar(
-                          content: Text('Document options coming soon!'),
-                        ),
-                        const SnackBar(content: Text('Document options coming soon!')),
-                      );
-                    },
-                  ),
-                ],
-              ),
-              const SizedBox(height: 8),
-              Text(
-                'Type: ${document.type.isNotEmpty ? document.type : 'N/A'}',
-                style: TextStyle(fontSize: 14, color: Colors.grey.shade700),
-              ),
-              const SizedBox(height: 4),
-              Text(
-                'File: ${document.fileName.isNotEmpty ? document.fileName : 'N/A'}',
-                style: TextStyle(fontSize: 14, color: Colors.grey.shade700),
-              ),
-              const SizedBox(height: 4),
-              Text(
-                'Expires: ${document.expiryDate.day.toString().padLeft(2, '0')}/${document.expiryDate.month.toString().padLeft(2, '0')}/${document.expiryDate.year}',
-                style: TextStyle(
-                  fontSize: 14,
-                  color:
-                      document.expiryDate.isBefore(DateTime.now())
-                          ? Colors.red
-                          : Colors.green.shade700,
-                  color: document.expiryDate.isBefore(DateTime.now()) ? Colors.red : Colors.green.shade700,
-                  fontWeight: FontWeight.w500,
-                ),
-              ),
-              const SizedBox(height: 12),
-              Align(
-                alignment: Alignment.bottomRight,
-                child: TextButton.icon(
-                  onPressed: () {
-                    ScaffoldMessenger.of(context).showSnackBar(
-                      SnackBar(
-                        content: Text('Viewing ${document.name} coming soon!'),
-                      ),
-                      SnackBar(content: Text('Viewing ${document.name} coming soon!')),
-                    );
-                  },
-                  icon: const Icon(Icons.visibility),
-                  label: const Text('View Document'),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
